Better class design
Instead of sharing the Queue<string> _packetsToSend variable, I put the
method within the class so as to keep _packetsToSend private to the
sending class, and have the class handle notifying the sending thread
that a new packet is ready to be sent, instead of the calling class
doing this.

Additionally, the performance improvement I strategized creating cannot
be implemented due to the fact that java's BufferedReader has a 8kB
character buffer size, and I speculate that, to implement the
improvements I wanted, I would have to dramatically increase this
default value, potentially causing a significant performance hit,
defeating the entire original purpose.